Server/web/requirements.txt
tensorflow==1.3.0
Flask==0.12.2
grpcio
tensorflow-serving-api
nltk
gunicorn==19.6.0